+++ src/cborx/__init__.py
@@ -6,7 +6,7 @@ from .encoder import CBOREncoder, DefaultCallback, dump, dumps
 from .exceptions import CBORDecodeError, CBOREncodeError, CBORError
 from .types import CBORSimpleValue, CBORTag, UndefinedType, undefined
 
-__version__ = "0.2.4"
+__version__ = "0.2.5"
 
 __all__ = [
     "CBORDecodeError",